+++ plot_data/core.py
@@ -160,6 +160,14 @@ class ReferencedObject(PlotDataObject):
         super().__init__(type_=type_, name=name)
 
 
+class Shape(ReferencedObject):
+    """ Shape object. """
+
+    def __init__(self, type_: str, reference_path: str = "#", tooltip: str = None, name: str = ""):
+        self.tooltip = tooltip
+        super().__init__(type_=type_, reference_path=reference_path, name=name)
+
+
 class Sample(ReferencedObject):
     """ Graph Point. """
 
@@ -426,7 +434,7 @@ class PointFamily(PlotDataObject):
         PlotDataObject.__init__(self, type_=None, name=name)
 
 
-class Text(ReferencedObject):
+class Text(Shape):
     """
     A class for displaying texts on canvas. Text is a primitive and can be instantiated by PrimitiveGroup.
 
@@ -450,7 +458,7 @@ class Text(ReferencedObject):
 
     def __init__(self, comment: str, position_x: float, position_y: float, text_style: TextStyle = None,
                  text_scaling: bool = None, max_width: float = None, height: float = None, multi_lines: bool = True,
-                 reference_path: str = "#", name: str = ''):
+                 reference_path: str = "#", tooltip: str = None, name: str = ''):
         self.comment = comment
         self.text_style = text_style
         self.position_x = position_x
@@ -459,7 +467,7 @@ class Text(ReferencedObject):
         self.max_width = max_width
         self.height = height
         self.multi_lines = multi_lines
-        super().__init__(type_='text', reference_path=reference_path, name=name)
+        super().__init__(type_='text', reference_path=reference_path, tooltip=tooltip, name=name)
 
     def mpl_plot(self, ax=None, color='k', alpha=1., **kwargs):
         """ Plots using Matplotlib. """
@@ -469,7 +477,7 @@ class Text(ReferencedObject):
         return ax
 
 
-class Line2D(ReferencedObject):
+class Line2D(Shape):
     """
     An infinite line. Line2D is a primitive and can be instantiated by PrimitiveGroups.
 
@@ -482,12 +490,12 @@ class Line2D(ReferencedObject):
     """
 
     def __init__(self, point1: List[float], point2: List[float], edge_style: EdgeStyle = None,
-                 reference_path: str = "#", name: str = ''):
+                 reference_path: str = "#", tooltip: str = None, name: str = ''):
         self.data = point1 + point2  # Retrocompatibility
         self.point1 = point1
         self.point2 = point2
         self.edge_style = edge_style
-        super().__init__(type_='line2d', reference_path=reference_path, name=name)
+        super().__init__(type_='line2d', reference_path=reference_path, tooltip=tooltip, name=name)
 
     def mpl_plot(self, ax=None, edge_style=None, **kwargs):
         """ Plots using matplotlib. """
@@ -505,7 +513,7 @@ class Line2D(ReferencedObject):
         return ax
 
 
-class LineSegment2D(ReferencedObject):
+class LineSegment2D(Shape):
     """
     A line segment. This is a primitive that can be called by PrimitiveGroup.
 
@@ -518,7 +526,7 @@ class LineSegment2D(ReferencedObject):
     """
 
     def __init__(self, point1: List[float], point2: List[float], edge_style: EdgeStyle = None,
-                 reference_path: str = "#", name: str = ''):
+                 reference_path: str = "#", tooltip: str = None, name: str = ''):
         # Data is used in typescript
         self.data = point1 + point2
         self.point1 = point1
@@ -528,7 +536,7 @@ class LineSegment2D(ReferencedObject):
             self.edge_style = EdgeStyle()
         else:
             self.edge_style = edge_style
-        super().__init__(type_='linesegment2d', reference_path=reference_path, name=name)
+        super().__init__(type_='linesegment2d', reference_path=reference_path, tooltip=tooltip, name=name)
 
     def bounding_box(self):
         """ Get 2D bounding box of current LineSegment2D. """
@@ -556,7 +564,7 @@ class LineSegment2D(ReferencedObject):
         return ax
 
 
-class Wire(ReferencedObject):
+class Wire(Shape):
     """
     A set of connected lines. It also provides highlighting feature.
 
@@ -572,8 +580,7 @@ class Wire(ReferencedObject):
                  reference_path: str = "#", name: str = ""):
         self.lines = lines
         self.edge_style = edge_style
-        self.tooltip = tooltip
-        super().__init__(type_="wire", reference_path=reference_path, name=name)
+        super().__init__(type_="wire", reference_path=reference_path, tooltip=tooltip, name=name)
 
     def mpl_plot(self, ax=None, **kwargs):
         """ Plots using matplotlib. """
@@ -588,7 +595,7 @@ class Wire(ReferencedObject):
         return ax
 
 
-class Circle2D(ReferencedObject):
+class Circle2D(Shape):
     """
     A circle. It is a primitive and can be instantiated by PrimitiveGroup.
 
@@ -613,8 +620,7 @@ class Circle2D(ReferencedObject):
         self.r = r
         self.cx = cx
         self.cy = cy
-        self.tooltip = tooltip
-        super().__init__(type_='circle', reference_path=reference_path, name=name)
+        super().__init__(type_='circle', reference_path=reference_path, tooltip=tooltip, name=name)
 
     def bounding_box(self):
         """ Get 2D bounding box of current Circle2D. """
@@ -644,7 +650,7 @@ class Circle2D(ReferencedObject):
         return ax
 
 
-class Rectangle(ReferencedObject):
+class Rectangle(Shape):
     """ Class to draw a rectangle. """
 
     def __init__(self, x_coord: float, y_coord: float, width: float, height: float, edge_style: EdgeStyle = None,
@@ -655,8 +661,7 @@ class Rectangle(ReferencedObject):
         self.height = height
         self.surface_style = surface_style
         self.edge_style = edge_style
-        self.tooltip = tooltip
-        super().__init__(type_='rectangle', reference_path=reference_path, name=name)
+        super().__init__(type_='rectangle', reference_path=reference_path, tooltip=tooltip, name=name)
 
     def bounding_box(self):
         """ Get 2D bounding box of current Circle2D. """
@@ -698,7 +703,7 @@ class RoundRectangle(Rectangle):
         self.radius = radius
 
 
-class Point2D(ReferencedObject):
+class Point2D(Shape):
     """
     A class for instantiating a point.
 
@@ -710,11 +715,12 @@ class Point2D(ReferencedObject):
     :type point_style: PointStyle
     """
 
-    def __init__(self, cx: float, cy: float, point_style: PointStyle = None, reference_path: str = "#", name: str = ''):
+    def __init__(self, cx: float, cy: float, point_style: PointStyle = None, reference_path: str = "#",
+                 tooltip: str = None, name: str = ''):
         self.cx = cx
         self.cy = cy
         self.point_style = point_style
-        super().__init__(type_='point', reference_path=reference_path, name=name)
+        super().__init__(type_='point', reference_path=reference_path, tooltip=tooltip, name=name)
 
     def bounding_box(self):
         """ Get 2D bounding box of current Circle2D. """
@@ -1114,7 +1120,7 @@ class ScatterMatrix(Figure):
                 for row in sample_attributes for col in sample_attributes]
 
 
-class Arc2D(ReferencedObject):
+class Arc2D(Shape):
     """
     A class for drawing arcs. Arc2D is a primitive and can be instantiated by PrimitiveGroup. By default,
     the arc is drawn anticlockwise.
@@ -1140,7 +1146,7 @@ class Arc2D(ReferencedObject):
     """
 
     def __init__(self, cx: float, cy: float, r: float, start_angle: float, end_angle: float, clockwise: bool = None,
-                 edge_style: EdgeStyle = None, reference_path: str = "#", name: str = ''):
+                 edge_style: EdgeStyle = None, reference_path: str = "#", tooltip: str = None, name: str = ''):
         self.cx = cx
         self.cy = cy
         self.r = r
@@ -1148,7 +1154,7 @@ class Arc2D(ReferencedObject):
         self.end_angle = end_angle
         self.clockwise = clockwise
         self.edge_style = edge_style
-        super().__init__(type_='arc', reference_path=reference_path, name=name)
+        super().__init__(type_='arc', reference_path=reference_path, tooltip=tooltip, name=name)
 
     def bounding_box(self):
         """ Get 2D bounding box of current Circle2D. """
@@ -1182,7 +1188,7 @@ class Arc2D(ReferencedObject):
         return ax
 
 
-class Contour2D(ReferencedObject):
+class Contour2D(Shape):
     """
     A Contour2D is a closed polygon that is formed by multiple primitives.
 
@@ -1202,9 +1208,8 @@ class Contour2D(ReferencedObject):
         self.plot_data_primitives = plot_data_primitives
         self.edge_style = edge_style
         self.surface_style = surface_style
-        self.tooltip = tooltip
         self.is_filled = surface_style is not None
-        super().__init__(type_='contour', reference_path=reference_path, name=name)
+        super().__init__(type_='contour', reference_path=reference_path, tooltip=tooltip, name=name)
 
     def bounding_box(self):
         """ Get 2D bounding box of current Contour2D. """
@@ -1508,7 +1513,7 @@ def plot_data_path(local: bool = False, version: str = None):
     """ Get path of plot_data package to write it in html file of Figure to draw. """
     version, folder, filename = get_current_link(version=version)
     if local:
-        core_path = os.sep.join(os.getcwd().split(os.sep)[:-1] + [folder, filename])
+        core_path = os.sep.join(__file__.split(os.sep)[:-2] + [folder, filename])
         if os.path.isfile(core_path):
             return core_path.replace(" ", "%20")
         print(f'Local compiled {core_path} not found, fall back to CDN')